requirements.txt
# Execute: pip install -r requirements.txt

streamlit
pandas
langchain
langchain-google-genai
matplotlib
seaborn
tabulate
langchain-experimental
python-dotenv